perbaikan taampilan, menambahkan tranksaksi event

+++ controllers/admin.js
@@ -2,12 +2,19 @@ const path = require( 'path' )
 const User = require( '../models/user' )
 const Event = require( '../models/event' )
 const fileHelper = require( '../util/file' )
+const { validationResult } = require( 'express-validator' )
 const Comment = require( '../models/comment' )
 const Like = require( '../models/like' )
-const { DeletePostLike, UpdateArrayPost, DeleteArrayPost, PushArrayUserEvent, PullArrayUserEvent } = require( '../functions/function' )
+const { DeletePostLike, PushArrayUserEvent, PullArrayUserEvent } = require( '../functions/function' )
 const { DeletePostComment } = require( '../functions/function' )
 exports.getPostEvent = ( req, res, next ) =>
 {
+    let message = ''
+    if ( message.length > 0 ) {
+        message = message[ 0 ]
+    } else {
+        message = null
+    }
     res.render(
         'event/postEvent',
         {
@@ -15,6 +22,8 @@ exports.getPostEvent = ( req, res, next ) =>
             path: '/admin/add-event',
             user: req.user,
             editMode: false,
+            errorMessage: message,
+            validErrors: []
         }
     )
 }
@@ -22,32 +31,56 @@ exports.getPostEvent = ( req, res, next ) =>
 exports.postPostEvent = ( req, res, next ) =>
 {
     const nameEvent = req.body.nameEvent
-    const pictureEvent = req.file.path.replace( '\\', '/' )
     const Ondate = req.body.OnDate
     const deskripsi = req.body.deskripsi
     const tempat = req.body.tempat
     const id = req.user._id
-    const event = new Event( {
-        userId: req.user._id,
-        nameEvent: nameEvent,
-        ImageEvent: pictureEvent,
-        tempat: tempat,
-        Ondate: Ondate,
-        Deskripsi: deskripsi,
-        html: "",
-        // like: ,
-        comment: []
-    } )
+    let harga = req.body.harga
+    const kapasitas = req.body.Kapasitas
+    console.log( req.body.Category )
+    const errors = validationResult( req )
+    if ( !errors.isEmpty() ) {
+        return res.render( 'event/postEvent', {
+            pageTitle: 'add event',
+            path: '/admin/add-event',
+            user: req.user,
+            editMode: false,
+            errorMessage: errors.array()[ 0 ].msg,
+            validErrors: errors.array()
+        } )
+    }
+    else {
+        if ( req.body.Category == 'free' ) {
+            console.log( 'set value harga =0' )
+            harga = 0
+        }
+        const pictureEvent = req.file.path.replace( '\\', '/' )
+        const event = new Event( {
+            userId: req.user._id,
+            nameEvent: nameEvent,
+            ImageEvent: pictureEvent,
+            tempat: tempat,
+            Ondate: Ondate,
+            Deskripsi: deskripsi,
+            html: "",
+            // like: ,
+            comment: [],
+            Harga: harga,
+            Kapasitas: kapasitas
+        } )
+        event.save().then( event =>
+        {
+            { PushArrayUserEvent( id, event._id ) }
+            res.redirect( '/' )
+        } ).catch( err =>
+        {
+            console.log( err )
+            res.redirect( '/500' )
+        } )
+    }
 
-    event.save().then( event =>
-    {
-        { PushArrayUserEvent( id, event._id ) }
-        res.redirect( '/' )
-    } ).catch( err =>
-    {
-        console.log( err )
-        res.redirect( '/500' )
-    } )
+
+
 }
 exports.getEditEvent = ( req, res, next ) =>
 {
@@ -64,12 +97,19 @@ exports.getEditEvent = ( req, res, next ) =>
             if ( !event ) {
                 return res.redirect( '/profil' )
             }
+            let message = ''
+            if ( message.length > 0 ) {
+                message = message[ 0 ]
+            } else {
+                message = null
+            }
             res.render( 'event/postEvent', {
                 pageTitle: 'Edit Event',
                 path: 'edit-event/:eventId',
                 event: event,
                 user: user,
-                editMode: true,
+                editMode: true, errorMessage: message,
+                validErrors: []
             } )
         } )
 }
@@ -116,29 +156,52 @@ exports.postEditEvent = ( req, res ) =>
     const UpdateOndate = req.body.OnDate
     const Updatedeskripsi = req.body.deskripsi
     const Updatetempat = req.body.tempat
+    let Updateharga = req.body.harga
+    const Updatekapasitas = req.body.Kapasitas
     Event.findById( eventId ).then( event =>
     {
-        event.nameEvent = UpdatenameEvent
-        event.OnDate = UpdateOndate
-        event.tempat = Updatetempat
-        event.html = ''
-        event.Deskripsi = Updatedeskripsi
-        if ( pictureEvent ) {
-            fileHelper.deleteFile( event.ImageEvent )
-            event.ImageEvent = pictureEvent.path.replace( '\\', '/' )
-
+        const errors = validationResult( req )
+        if ( !errors.isEmpty() ) {
+            return res.render( 'event/postEvent', {
+                pageTitle: 'Edit Event',
+                path: 'edit-event/:eventId',
+                user: req.user,
+                editMode: false,
+                errorMessage: errors.array()[ 0 ].msg,
+                validErrors: errors.array()
+            } )
         }
-        return event.save().then( result =>
-        {
-            res.redirect( '/profil' )
-        } ).catch( err =>
-        {
-            console.log( err )
-            res.redirect( '/500' )
-        } )
-    } ).catch( err =>
-    {
-        console.log( err )
-        res.redirect( '/500' )
+        else {
+            if ( req.body.Category == 'free' ) {
+                console.log( 'set value harga =0' )
+                Updateharga = 0
+            }
+            event.Harga = Updateharga
+            event.Kapasitas = Updatekapasitas
+            event.nameEvent = UpdatenameEvent
+            event.OnDate = UpdateOndate
+            event.tempat = Updatetempat
+            event.html = ''
+            event.Deskripsi = Updatedeskripsi
+            if ( pictureEvent ) {
+                fileHelper.deleteFile( event.ImageEvent )
+                event.ImageEvent = pictureEvent.path.replace( '\\', '/' )
+
+            }
+            return event.save().then( result =>
+            {
+                res.redirect( '/profil' )
+            } ).catch( err =>
+            {
+                console.log( err )
+                res.redirect( '/500' )
+            } )
+                // } )
+                .catch( err =>
+                {
+                    console.log( err )
+                    res.redirect( '/500' )
+                } )
+        }
     } )
 }
